Develop

+++ CoolParking/CoolParking.WebAPI/Controllers/VehiclesController.cs
@@ -1,5 +1,7 @@
 ﻿using CoolParking.BL.Interfaces;
 using CoolParking.BL.Models;
+using CoolParking.Common.DTO;
+using CoolParking.Common.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CoolParking.WebAPI.Controllers
@@ -9,66 +11,68 @@ namespace CoolParking.WebAPI.Controllers
     public class VehiclesController : ControllerBase
     {
         private readonly IParkingService _parkingService;
-        public VehiclesController(IParkingService parkingService)
+        private readonly IVehicleService _vehicleService;
+        public VehiclesController(IParkingService parkingService, IVehicleService vehicleService)
         {
             _parkingService = parkingService;
+            _vehicleService = vehicleService;   
         }
 
         //api/vehicles
         [HttpGet]
-        public ActionResult<IEnumerable<Vehicle>> GetAll() //Tested
+        public ActionResult<IEnumerable<VehicleDTO>> GetAll() 
         {
-            return Ok(_parkingService.GetVehicles());
+            return Ok(_vehicleService.GetVehicles());
         }
 
         //api/vehicles/id 
         [HttpGet("{id}", Name = "GetById")]
-        public ActionResult<Vehicle> GetById(string id) //Tested
+        public ActionResult<VehicleDTO> GetById(string id) 
         {
             if (!Vehicle.IsValidId(id))
             {
                 return BadRequest();
             }
 
-            var vehicle = _parkingService.GetVehicles().FirstOrDefault(v => v.Id == id);
+            var vehicleDTO = _vehicleService.GetVehicleById(id);
 
-            if (vehicle == null)
+            if (vehicleDTO == null)
             {
                 return NotFound();
             }
 
-            return Ok(vehicle);
+            return Ok(vehicleDTO);
         }
 
         //api/vehicles
         [HttpPost]
-        public IActionResult Add([FromBody] Vehicle vehicle) //Tested
+        public IActionResult Add([FromBody] VehicleDTO vehicleDTO) 
         {
-            if (!Vehicle.IsValidId(vehicle.Id))
+            if (!Vehicle.IsValidId(vehicleDTO.Id) && vehicleDTO.Balance >= Settings.Tariffs[(int)vehicleDTO.VehicleType])
             {
                 return BadRequest();
             }
-            
-            _parkingService.AddVehicle(vehicle);
 
-            return CreatedAtRoute("GetById", new { id = vehicle.Id }, vehicle);
+            _vehicleService.AddVehicle(vehicleDTO);
+
+            return CreatedAtRoute("GetById", new { id = vehicleDTO.Id }, vehicleDTO);
         }
 
         //api/vehicles/id
         [HttpDelete("{id}")]
-        public IActionResult Delete(string id) //Tested
-        
+        public IActionResult Delete(string id) 
         {
-            if (!Vehicle.IsValidId(id))
-            {
-                return BadRequest();
-            }
 
             var vehicle = _parkingService.GetVehicles().FirstOrDefault(v => v.Id == id);
 
             if (vehicle == null)
             {
                 return NotFound();
+            }
+
+            if (!Vehicle.IsValidId(id) || vehicle.Balance < 0)
+            {
+                return BadRequest();
             }
 
             _parkingService.RemoveVehicle(id);
